requirements.txt
grpcio
grpcio-tools
numpy
pandas
scikit-learn
joblib
boto3
s3fs